requirements/prod.txt
# Production dependencies
-r base.txt  # Include all base requirements

# Security
django-secure==2.0.0
python-dotenv==1.0.0

# Performance
django-compressor==4.3.1
django-redis==5.2.0

# Monitoring
sentry-sdk==1.30.0
django-statsd-mozilla==0.4.0
